Improved PluginProvider.filter path matching
The previous matching method checked for a partial path ANYWHERE after the plugin directory of a TSP file path.  This works for most cases, but it falls over when encountering duplicates:

Imagine we have two command files "example.tsp" and "nested/example.tsp".  If we tried to filter on "example" (for the top-level command), then both commands would be matched.

Now the matcher strictly checks start of the path segment following the plugin directory name.